// src/modules/system/messages/exception-text/exception-text.enum.ts
export enum ExceptionText {
  BAD_REQUEST = 'Bad Request',
  CONFLICT = 'Conflict',
  FORBIDDEN = 'Forbidden',
  NOT_FOUND = 'Not Found',
  UNAUTHORIZED = 'Unauthorized',
  UNPROCESSABLE_ENTITY = 'Unprocessable Entity',
  UNPROCESSABLE_ENTITY_EXCEPTION = 'UnprocessableEntityException',
  UNPROCESSABLE__ENTITY__EXCEPTION = 'Unprocessable Entity Exception',
}